Split profile functionality from build_yaml
No changes - just move the code from one big file to separate ones.
Import the symbols to the original file for compatibility.
