+++ ssg/build_yaml.py
@@ -1,7 +1,6 @@
 from __future__ import absolute_import
 from __future__ import print_function
 
-from collections import defaultdict
 from copy import deepcopy
 import datetime
 import json
@@ -9,17 +8,14 @@ import os
 import os.path
 import re
 import sys
-from xml.sax.saxutils import escape
 import glob
 
 
 import ssg.build_remediations
 from .build_cpe import CPEDoesNotExist, CPEALLogicalTest, CPEALFactRef, ProductCPEs
 from .constants import (XCCDF12_NS,
-                        XCCDF_REFINABLE_PROPERTIES,
                         OSCAP_BENCHMARK,
                         OSCAP_GROUP,
-                        OSCAP_PROFILE,
                         OSCAP_RULE,
                         OSCAP_VALUE,
                         SCE_SYSTEM,
@@ -46,8 +42,18 @@ from .utils import required_key, mkdir_p
 
 from .xml import ElementTree as ET, add_xhtml_namespace, register_namespaces, parse_file
 from .shims import unicode_func
-from .entities.common import XCCDFEntity
 import ssg.build_stig
+
+from .entities.common import (
+    XCCDFEntity,
+    SelectionHandler,
+    add_sub_element,
+    derive_id_from_file_name
+)
+from .entities.profile import (
+    Profile,
+    ProfileWithInlinePolicies,
+)
 
 
 def add_sub_element(parent, tag, ns, data):
@@ -180,410 +186,6 @@ def add_benchmark_metadata(element, contributors_file):
 
     source = ET.SubElement(metadata, "{%s}source" % dc_namespace)
     source.text = SSG_BENCHMARK_LATEST_URI
-
-
-class SelectionHandler(object):
-    def __init__(self):
-        self.refine_rules = defaultdict(list)
-        self.variables = dict()
-        self.unselected = []
-        self.unselected_groups = []
-        self.selected = []
-
-    @property
-    def selections(self):
-        selections = []
-        for item in self.selected:
-            selections.append(str(item))
-        for item in self.unselected:
-            selections.append("!"+str(item))
-        for varname in self.variables.keys():
-            selections.append(varname+"="+self.variables.get(varname))
-        for rule, refinements in self.refine_rules.items():
-            for prop, val in refinements:
-                selections.append("{rule}.{property}={value}"
-                                  .format(rule=rule, property=prop, value=val))
-        return selections
-
-    @selections.setter
-    def selections(self, entries):
-        for item in entries:
-            self.apply_selection(item)
-
-    def apply_selection(self, item):
-        if "." in item:
-            rule, refinement = item.split(".", 1)
-            property_, value = refinement.split("=", 1)
-            if property_ not in XCCDF_REFINABLE_PROPERTIES:
-                msg = ("Property '{property_}' cannot be refined. "
-                       "Rule properties that can be refined are {refinables}. "
-                       "Fix refinement '{rule_id}.{property_}={value}' in profile '{profile}'."
-                       .format(property_=property_, refinables=XCCDF_REFINABLE_PROPERTIES,
-                               rule_id=rule, value=value, profile=self.id_)
-                       )
-                raise ValueError(msg)
-            self.refine_rules[rule].append((property_, value))
-        elif "=" in item:
-            varname, value = item.split("=", 1)
-            self.variables[varname] = value
-        elif item.startswith("!"):
-            self.unselected.append(item[1:])
-        else:
-            self.selected.append(item)
-
-    def _subtract_refinements(self, extended_refinements):
-        """
-        Given a dict of rule refinements from the extended profile,
-        "undo" every refinement prefixed with '!' in this profile.
-        """
-        for rule, refinements in list(self.refine_rules.items()):
-            if rule.startswith("!"):
-                for prop, val in refinements:
-                    extended_refinements[rule[1:]].remove((prop, val))
-                del self.refine_rules[rule]
-        return extended_refinements
-
-    def update_with(self, rhs):
-        extended_selects = set(rhs.selected)
-        extra_selections = extended_selects.difference(set(self.selected))
-        self.selected.extend(list(extra_selections))
-
-        updated_variables = dict(rhs.variables)
-        updated_variables.update(self.variables)
-        self.variables = updated_variables
-
-        extended_refinements = deepcopy(rhs.refine_rules)
-        updated_refinements = self._subtract_refinements(extended_refinements)
-        updated_refinements.update(self.refine_rules)
-        self.refine_rules = updated_refinements
-
-
-class Profile(XCCDFEntity, SelectionHandler):
-    """Represents XCCDF profile
-    """
-    KEYS = dict(
-        title=lambda: "",
-        description=lambda: "",
-        extends=lambda: "",
-        metadata=lambda: None,
-        reference=lambda: None,
-        selections=lambda: list(),
-        unselected_groups=lambda: list(),
-        platforms=lambda: set(),
-        cpe_names=lambda: set(),
-        platform=lambda: None,
-        filter_rules=lambda: "",
-        ** XCCDFEntity.KEYS
-    )
-
-    MANDATORY_KEYS = {
-        "title",
-        "description",
-        "selections",
-    }
-
-    @classmethod
-    def process_input_dict(cls, input_contents, env_yaml, product_cpes):
-        input_contents = super(Profile, cls).process_input_dict(input_contents, env_yaml)
-
-        platform = input_contents.get("platform")
-        if platform is not None:
-            input_contents["platforms"].add(platform)
-
-        if env_yaml:
-            for platform in input_contents["platforms"]:
-                try:
-                    new_cpe_name = product_cpes.get_cpe_name(platform)
-                    input_contents["cpe_names"].add(new_cpe_name)
-                except CPEDoesNotExist:
-                    msg = (
-                        "Unsupported platform '{platform}' in a profile."
-                        .format(platform=platform))
-                    raise CPEDoesNotExist(msg)
-
-        return input_contents
-
-    @property
-    def rule_filter(self):
-        if self.filter_rules:
-            return rule_filter_from_def(self.filter_rules)
-        else:
-            return noop_rule_filterfunc
-
-    def _add_selects(self, element, selections, prefix, selected):
-        for selection in selections:
-            select = ET.Element("{%s}select" % XCCDF12_NS)
-            select.set("idref", prefix + selection)
-            select.set("selected", selected)
-            element.append(select)
-
-    def to_xml_element(self):
-        element = ET.Element('{%s}Profile' % XCCDF12_NS)
-        element.set("id", OSCAP_PROFILE + self.id_)
-        if self.extends:
-            element.set("extends", self.extends)
-        title = add_sub_element(element, "title", XCCDF12_NS, self.title)
-        title.set("override", "true")
-        desc = add_sub_element(
-            element, "description", XCCDF12_NS, self.description)
-        desc.set("override", "true")
-
-        if self.reference:
-            add_sub_element(
-                element, "reference", XCCDF12_NS, escape(self.reference))
-
-        for cpe_name in self.cpe_names:
-            plat = ET.SubElement(element, "{%s}platform" % XCCDF12_NS)
-            plat.set("idref", cpe_name)
-
-        self._add_selects(element, self.selected, OSCAP_RULE, "true")
-        self._add_selects(element, self.unselected, OSCAP_RULE, "false")
-        self._add_selects(element, self.unselected_groups, OSCAP_GROUP, "false")
-
-        for value_id, selector in self.variables.items():
-            refine_value = ET.Element("{%s}refine-value" % XCCDF12_NS)
-            refine_value.set("idref", OSCAP_VALUE + value_id)
-            refine_value.set("selector", selector)
-            element.append(refine_value)
-
-        for refined_rule, refinement_list in self.refine_rules.items():
-            refine_rule = ET.Element("{%s}refine-rule" % XCCDF12_NS)
-            refine_rule.set("idref", OSCAP_RULE + refined_rule)
-            for refinement in refinement_list:
-                refine_rule.set(refinement[0], refinement[1])
-            element.append(refine_rule)
-
-        return element
-
-    def get_rule_selectors(self):
-        return self.selected + self.unselected
-
-    def get_variable_selectors(self):
-        return self.variables
-
-    def validate_refine_rules(self, rules):
-        existing_rule_ids = [r.id_ for r in rules]
-        for refine_rule, refinement_list in self.refine_rules.items():
-            # Take first refinement to ilustrate where the error is
-            # all refinements in list are invalid, so it doesn't really matter
-            a_refinement = refinement_list[0]
-
-            if refine_rule not in existing_rule_ids:
-                msg = (
-                    "You are trying to refine a rule that doesn't exist. "
-                    "Rule '{rule_id}' was not found in the benchmark. "
-                    "Please check all rule refinements for rule: '{rule_id}', for example: "
-                    "- {rule_id}.{property_}={value}' in profile {profile_id}."
-                    .format(rule_id=refine_rule, profile_id=self.id_,
-                            property_=a_refinement[0], value=a_refinement[1])
-                    )
-                raise ValueError(msg)
-
-            if refine_rule not in self.get_rule_selectors():
-                msg = ("- {rule_id}.{property_}={value}' in profile '{profile_id}' is refining "
-                       "a rule that is not selected by it. The refinement will not have any "
-                       "noticeable effect. Either select the rule or remove the rule refinement."
-                       .format(rule_id=refine_rule, property_=a_refinement[0],
-                               value=a_refinement[1], profile_id=self.id_)
-                       )
-                raise ValueError(msg)
-
-    def validate_variables(self, variables):
-        variables_by_id = dict()
-        for var in variables:
-            variables_by_id[var.id_] = var
-
-        for var_id, our_val in self.variables.items():
-            if var_id not in variables_by_id:
-                all_vars_list = [" - %s" % v for v in variables_by_id.keys()]
-                msg = (
-                    "Value '{var_id}' in profile '{profile_name}' is not known. "
-                    "We know only variables:\n{var_names}"
-                    .format(
-                        var_id=var_id, profile_name=self.id_,
-                        var_names="\n".join(sorted(all_vars_list)))
-                )
-                raise ValueError(msg)
-
-            allowed_selectors = [str(s) for s in variables_by_id[var_id].options.keys()]
-            if our_val not in allowed_selectors:
-                msg = (
-                    "Value '{var_id}' in profile '{profile_name}' "
-                    "uses the selector '{our_val}'. "
-                    "This is not possible, as only selectors {all_selectors} are available. "
-                    "Either change the selector used in the profile, or "
-                    "add the selector-value pair to the variable definition."
-                    .format(
-                        var_id=var_id, profile_name=self.id_, our_val=our_val,
-                        all_selectors=allowed_selectors,
-                    )
-                )
-                raise ValueError(msg)
-
-    def validate_rules(self, rules, groups):
-        existing_rule_ids = [r.id_ for r in rules]
-        rule_selectors = self.get_rule_selectors()
-        for id_ in rule_selectors:
-            if id_ in groups:
-                msg = (
-                    "You have selected a group '{group_id}' instead of a "
-                    "rule. Groups have no effect in the profile and are not "
-                    "allowed to be selected. Please remove '{group_id}' "
-                    "from profile '{profile_id}' before proceeding."
-                    .format(group_id=id_, profile_id=self.id_)
-                )
-                raise ValueError(msg)
-            if id_ not in existing_rule_ids:
-                msg = (
-                    "Rule '{rule_id}' was not found in the benchmark. Please "
-                    "remove rule '{rule_id}' from profile '{profile_id}' "
-                    "before proceeding."
-                    .format(rule_id=id_, profile_id=self.id_)
-                )
-                raise ValueError(msg)
-
-    def _find_empty_groups(self, group, profile_rules):
-        is_empty = True
-        empty_groups = []
-        for child in group.groups.values():
-            child_empty, child_empty_groups = self._find_empty_groups(child, profile_rules)
-            if not child_empty:
-                is_empty = False
-            empty_groups.extend(child_empty_groups)
-        if is_empty:
-            group_rules = set(group.rules.keys())
-            if profile_rules & group_rules:
-                is_empty = False
-        if is_empty:
-            empty_groups.append(group.id_)
-        return is_empty, empty_groups
-
-    def unselect_empty_groups(self, root_group):
-        # Unselecting empty groups is necessary to make HTML guides shorter
-        # and the XCCDF more usable in tools such as SCAP Workbench.
-        profile_rules = set(self.selected)
-        is_empty, empty_groups = self._find_empty_groups(root_group, profile_rules)
-        if is_empty:
-            msg = "Profile {0} unselects all groups.".format(self.id_)
-            raise ValueError(msg)
-        self.unselected_groups.extend(sorted(empty_groups))
-
-    def __sub__(self, other):
-        profile = Profile(self.id_)
-        profile.title = self.title
-        profile.description = self.description
-        profile.extends = self.extends
-        profile.platforms = self.platforms
-        profile.platform = self.platform
-        profile.selected = list(set(self.selected) - set(other.selected))
-        profile.selected.sort()
-        profile.unselected = list(set(self.unselected) - set(other.unselected))
-        profile.variables = dict((k, v) for (k, v) in self.variables.items()
-                             if k not in other.variables or v != other.variables[k])
-        return profile
-
-
-class ResolvableProfile(Profile):
-    def __init__(self, * args, ** kwargs):
-        super(ResolvableProfile, self).__init__(* args, ** kwargs)
-        self.resolved = False
-
-    def _controls_ids_to_controls(self, controls_manager, policy_id, control_id_list):
-        items = [controls_manager.get_control(policy_id, cid) for cid in control_id_list]
-        return items
-
-    def resolve_controls(self, controls_manager):
-        pass
-
-    def extend_by(self, extended_profile):
-        self.update_with(extended_profile)
-
-    def resolve_selections_with_rules(self, rules_by_id):
-        selections = set()
-        for rid in self.selected:
-            if rid not in rules_by_id:
-                continue
-            rule = rules_by_id[rid]
-            if not self.rule_filter(rule):
-                continue
-            selections.add(rid)
-        self.selected = list(selections)
-
-    def resolve(self, all_profiles, rules_by_id, controls_manager=None):
-        if self.resolved:
-            return
-
-        if controls_manager:
-            self.resolve_controls(controls_manager)
-
-        self.resolve_selections_with_rules(rules_by_id)
-
-        if self.extends:
-            if self.extends not in all_profiles:
-                msg = (
-                    "Profile {name} extends profile {extended}, but "
-                    "only profiles {known_profiles} are available for resolution."
-                    .format(name=self.id_, extended=self.extends,
-                            known_profiles=list(all_profiles.keys())))
-                raise RuntimeError(msg)
-            extended_profile = all_profiles[self.extends]
-            extended_profile.resolve(all_profiles, rules_by_id, controls_manager)
-
-            self.extend_by(extended_profile)
-
-        self.selected = [s for s in set(self.selected) if s not in self.unselected]
-
-        self.unselected = []
-        self.extends = None
-
-        self.selected = sorted(self.selected)
-
-        for rid in self.selected:
-            if rid not in rules_by_id:
-                msg = (
-                    "Rule {rid} is selected by {profile}, but the rule is not available. "
-                    "This may be caused by a discrepancy of prodtypes."
-                    .format(rid=rid, profile=self.id_))
-                raise ValueError(msg)
-
-        self.resolved = True
-
-
-class ProfileWithInlinePolicies(ResolvableProfile):
-    def __init__(self, * args, ** kwargs):
-        super(ProfileWithInlinePolicies, self).__init__(* args, ** kwargs)
-        self.controls_by_policy = defaultdict(list)
-
-    def apply_selection(self, item):
-        # ":" is the delimiter for controls but not when the item is a variable
-        if ":" in item and "=" not in item:
-            policy_id, control_id = item.split(":", 1)
-            self.controls_by_policy[policy_id].append(control_id)
-        else:
-            super(ProfileWithInlinePolicies, self).apply_selection(item)
-
-    def _process_controls_ids_into_controls(self, controls_manager, policy_id, controls_ids):
-        controls = []
-        for cid in controls_ids:
-            if not cid.startswith("all"):
-                controls.extend(
-                    self._controls_ids_to_controls(controls_manager, policy_id, [cid]))
-            elif ":" in cid:
-                _, level_id = cid.split(":", 1)
-                controls.extend(
-                    controls_manager.get_all_controls_of_level(policy_id, level_id))
-            else:
-                controls.extend(
-                    controls_manager.get_all_controls(policy_id))
-        return controls
-
-    def resolve_controls(self, controls_manager):
-        for policy_id, controls_ids in self.controls_by_policy.items():
-            controls = self._process_controls_ids_into_controls(
-                controls_manager, policy_id, controls_ids)
-
-            for c in controls:
-                self.update_with(c)
 
 
 class Value(XCCDFEntity):
@@ -2071,7 +1673,3 @@ def add_platform_if_not_defined(platform, product_cpes):
             return p
     product_cpes.platforms[platform.id_] = platform
     return platform
-
-
-def derive_id_from_file_name(filename):
-    return os.path.splitext(filename)[0]
